Agrego Feature: tests para OrdenDomainService

+++ SOI.Domain/Services/IOrdenDomainService.cs
@@ -3,5 +3,5 @@
 public interface IOrdenDomainService
 {
     public void Validar(int tipoActivoId, decimal? precio);
-    public decimal CalcularMontoTotal(int tipoActivoId, decimal precioActivoBD, decimal? precioRequest, decimal cantidad);
+    public decimal CalcularMontoTotal(int tipoActivoId, decimal precioActivoBD, decimal? precioRequest, int cantidad);
 }
